+++ node/packages/core/src/types.ts
@@ -1,4 +1,5 @@
 import { CompletionMeta } from "@instructor-ai/instructor";
+import { JSONContent } from "@tiptap/core";
 import { Edge, Node as XYFlowNode, Position } from "@xyflow/react";
 import { ChatCompletionMessageParam } from "openai/resources/index.mjs";
 import { z } from "zod";
@@ -13,6 +14,7 @@ export enum NodeType {
   ResponseSingle = "response_single",
   ResponseMultiple = "response_multiple",
   ResponseTable = "response_table",
+  ResponseMarkMap = "response_markmap",
   PaperFinder = "paper_finder",
   PaperQA = "paper_qa",
   ExternalData = "external_data",
@@ -28,6 +30,7 @@ export const nodeTypeMap = {
   [NodeType.ResponseTable]: "Response (table)",
   [NodeType.ResponseSingle]: "Responses (one node)",
   [NodeType.ResponseMultiple]: "Responses (many nodes)",
+  [NodeType.ResponseMarkMap]: "Responses (MarkMap)",
   [NodeType.PaperFinder]: "Paper Finder",
   [NodeType.PaperQA]: "Paper QA",
   [NodeType.ExternalData]: "External Data",
@@ -36,6 +39,12 @@ export const nodeTypeMap = {
 export type SpaceNode = {
   id: string;
   title: string;
+};
+
+export type SourceNode = {
+  id: string;
+  spaceId?: string;
+  nodeId?: string;
 };
 
 export type CanvasNode = XYFlowNode<
@@ -56,6 +65,7 @@ export type CanvasNode = XYFlowNode<
       spaceId: string;
       depth: number;
     };
+    sourceNode?: SourceNode;
   },
   "GraphNode" | "ExternalNode"
 >;
@@ -68,11 +78,34 @@ export interface Canvas {
   topologicallySortedNodes: string[];
 }
 
+export type ExportNodeSingle = {
+  id: string;
+  width: number | null | undefined;
+  height: number | null | undefined;
+  type?: string;
+  title: string;
+  position: {
+    x: number;
+    y: number;
+  };
+  data?: {
+    borderColor?: string;
+    type?: NodeType;
+  };
+  content?: JSONContent;
+};
+
+export type ExportNode = ExportNodeSingle & {
+  nodes: ExportNodeSingle[];
+  edges: CanvasEdge[];
+};
+
 export interface Task {
   inputNodes: CanvasNode[];
   outputNode: CanvasNode | null;
   promptNode: CanvasNode;
   loop?: boolean;
+  sourceNodeInfo?: SourceNode;
 }
 
 export interface ExecutionContext {